more code extracting
