+++ src/components/Heatmap/V2.tsx
@@ -9,6 +9,7 @@ import { RootState } from "state"
 import { Colors } from "consts/Colors"
 import { Spacer } from "components/Spacer"
 
+import { StatusBars } from "./StatusBars"
 import {
   Tooltip,
   Total,
@@ -330,48 +331,7 @@ export default function Heatmap({ data }: Props) {
       <Tooltip ref={tooltipRef}>
         &nbsp;-&nbsp;&nbsp;-&nbsp;&nbsp;&nbsp;&nbsp;
       </Tooltip>
-      {[...springs].reverse().map((props, i) => {
-        const grade = refs.length - 1 - i
-        return (
-          <FlexRowAlignCenter key={i}>
-            <VGrade>
-              <small>V{grade}</small>
-            </VGrade>
-            <Spacer x={10} />
-            <animated.div // container for percentage bars
-              style={{
-                display: `flex`,
-                width: props.total.to((a) => (a + 2) * 30),
-              }}
-            >
-              <Total style={{ width: props.totalWidth }}>
-                <Sends style={{ width: props.sendWidth, ...props }}></Sends>
-              </Total>
-              <Spacer x={5} />
-              <small style={{ minWidth: 60 }} ref={refs[i]}>
-                ?
-              </small>
-            </animated.div>
-          </FlexRowAlignCenter>
-        )
-      })}
+      <StatusBars springs={springs} refs={refs} />
     </>
   )
 }
-
-const FlexRowAlignCenter = styled.div`
-  display: flex;
-  align-items: center;
-`
-
-const VGrade = styled(animated.div)`
-  width: 30px;
-  height: 30px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  border: 1px dotted ${(p: BaseProps) => p.theme.commentRenderer.borderColor};
-  border-radius: 50%;
-  margin-top: 2px;
-  margin-bottom: 2px;
-`
